Move VARS and RUN to folders

+++ config.sh
@@ -1,43 +1,54 @@
 #!/bin/bash
+#$1 - vars file
+#$2 - log file
 
+source $1
 source ./general.sh
-source ./vars.sh
 
 clear
 
+exec > >(tee $2) 2>&1
+
 echo "Ubuntu 18.04 config started..."
-echo "Set hostname..."
-echo "========================================================================================"
-sudo hostnamectl set-hostname $HostName
-echo $HostName
-echo ""
-echo "Set NTP server..."
-echo "========================================================================================"
-echo "NTP=$NtpServer" >> /etc/systemd/timesyncd.conf
-echo $NtpServer
-timedatectl set-ntp true
-service systemd-timesyncd  restart
-service systemd-timesyncd status
-sleep 5s
-echo ""
+if [ -n $HostName ]; then
+	echo "Set hostname..."
+	echo "========================================================================================"
+	sudo hostnamectl set-hostname $HostName
+	echo $HostName
+	echo ""
+fi
+if [ -n $NtpServer ]; then
+	echo "Set NTP server..."
+	echo "========================================================================================"
+	echo "NTP=$NtpServer" >> /etc/systemd/timesyncd.conf
+	echo $NtpServer
+	timedatectl set-ntp true
+	service systemd-timesyncd  restart
+	service systemd-timesyncd status
+	sleep 5s
+	echo ""
+fi
 echo "Packets update..."
 echo "========================================================================================"
 	sudo apt-get update && apt-get upgrade -y
 echo ""
 sleep 2s
-echo "Time zone and Locale Configuring..."
-echo "========================================================================================"
-	sudo ln -fs /usr/share/zoneinfo/$Timezone /etc/localtime \
-	&& dpkg-reconfigure -f noninteractive tzdata
-	sed -i -e 's/# en_US.UTF-8 UTF-8/en_US.UTF-8 UTF-8/' /etc/locale.gen \
-	&& locale-gen
-echo ""
-sleep 2s
-echo "SSHD Configuring..."
-echo "========================================================================================"
+if [ -n $Timezone ]; then
+	echo "Time zone and Locale Configuring..."
+	echo "========================================================================================"
+		sudo ln -fs /usr/share/zoneinfo/$Timezone /etc/localtime \
+		&& dpkg-reconfigure -f noninteractive tzdata
+		sed -i -e 's/# en_US.UTF-8 UTF-8/en_US.UTF-8 UTF-8/' /etc/locale.gen \
+		&& locale-gen
+	sleep 2s
+	echo ""
+fi
+if [ -n $sshdfile ]; then
+	echo "SSHD Configuring..."
+	echo "========================================================================================"
 	file=$rootpath$sshdfile
 	if [ -f $file ]; then
-        echo "file $file - exist!"
+		echo "file $file - exist!"
 		AddOrReplaceParamInFile "# Disable root login for ssh" "" $file 
 		AddOrReplaceParamInFile "PermitRootLogin " "no"  $file $FirstLevel
 		AddOrReplaceParamInFile "ChallengeResponseAuthentication " "no" $file $FirstLevel
@@ -74,15 +85,17 @@ echo "==========================================================================
 		AddOrReplaceParamInFile "ListenAddress " "0.0.0.0" $file $FirstLevel
 		RemoveEmptyStrings $file
 	else
-        echo "file $file - NOT exist!!!!"
-    fi	
-echo ""
-sleep 2s
-echo "KERNEL Configuring..."
-echo "========================================================================================"
+		echo "file $file - NOT exist!!!!"
+	fi	
+	sleep 2s
+	echo ""
+fi
+if [ -n $sysctlfile ]; then
+	echo "KERNEL Configuring..."
+	echo "========================================================================================"
 	file=$rootpath$sysctlfile
 	if [ -f $file ]; then
-        echo "file $file - exist!"
+		echo "file $file - exist!"
 		AddOrReplaceParamInFile "# Turn on execshield" ""  $file
 		#AddOrReplaceParamInFile "kernel.exec-shield" "=1"  $file $FirstLevel #options turned off in Ubuntu 18
 		AddOrReplaceParamInFile "kernel.randomize_va_space" "=1"  $file $FirstLevel
@@ -99,62 +112,68 @@ echo "==========================================================================
 		AddOrReplaceParamInFile "net.ipv4.tcp_syncookies" "=1"  $file $FirstLevel
 		RemoveEmptyStrings $file
 	else
-        echo "file $file - NOT exist!!!!"
-    fi
-echo ""
-sleep 2s
-echo "Disable IPV6..."
-echo "========================================================================================"
+		echo "file $file - NOT exist!!!!"
+	fi
+	sleep 2s
+	echo ""	
+fi
+if [ -n $grubfile ]; then
+	echo "Disable IPV6..."
+	echo "========================================================================================"
 	file=$rootpath$grubfile
 	if [ -f $file ]; then
-        echo "file $file - exist!"
+		echo "file $file - exist!"
 		AddOrReplaceParamInFile "# Disable IPV6" ""  $file
 		AddOrReplaceParamInFile "GRUB_CMDLINE_LINUX_DEFAULT=" '"ipv6.disable=1"'  $file $FirstLevel
 		AddOrReplaceParamInFile "GRUB_CMDLINE_LINUX=" '"ipv6.disable=1"'  $file $FirstLevel
 		RemoveEmptyStrings $file
 	else
-        echo "file $file - NOT exist!!!!"
-    fi	
-echo ""
-sleep 2s
-echo "Disable USB/firewire/thunderbolt devices..."
-echo "========================================================================================"
-	file=$rootpath$usbfile
-	if [ -f $file ]; then
-        echo "file $file - exist!"
-		AddOrReplaceParamInFile "# Disable USB devices" ""  $file
-		AddOrReplaceParamInFile "install usb-storage " "/bin/true"  $file $FirstLevel
-		RemoveEmptyStrings $file
-	else
-        echo "file $file - NOT exist!!!!"
-    fi
+		echo "file $file - NOT exist!!!!"
+	fi	
+	sleep 2s
+	echo ""
+fi
+if [ -n $usbfile ]; then
+	echo "Disable USB/firewire/thunderbolt devices..."
+	echo "========================================================================================"
+		file=$rootpath$usbfile
+		if [ -f $file ]; then
+			echo "file $file - exist!"
+			AddOrReplaceParamInFile "# Disable USB devices" ""  $file
+			AddOrReplaceParamInFile "install usb-storage " "/bin/true"  $file $FirstLevel
+			RemoveEmptyStrings $file
+		else
+			echo "file $file - NOT exist!!!!"
+		fi
 
-	file=$rootpath$firewirefile
-	if [ -f $file ]; then
-        echo "file $file - exist!"
-		AddOrReplaceParamInFile "# Disable firewire devices" ""  $file
-		AddOrReplaceParamInFile "blacklist firewire-core" ""  $file $FirstLevel
-		RemoveEmptyStrings $file
-	else
-        echo "file $file - NOT exist!!!!"
-    fi
+		file=$rootpath$firewirefile
+		if [ -f $file ]; then
+			echo "file $file - exist!"
+			AddOrReplaceParamInFile "# Disable firewire devices" ""  $file
+			AddOrReplaceParamInFile "blacklist firewire-core" ""  $file $FirstLevel
+			RemoveEmptyStrings $file
+		else
+			echo "file $file - NOT exist!!!!"
+		fi
 
-	file=$rootpath$thunderboltfile
-	if [ -f $file ]; then
-        echo "file $file - exist!"
-		AddOrReplaceParamInFile "# Disable thunderbolt devices" ""  $file
-		AddOrReplaceParamInFile "blacklist thunderbolt" ""  $file $FirstLevel
-		RemoveEmptyStrings $file
-	else
-        echo "file $file - NOT exist!!!!"
-    fi
-echo ""
-echo "Config Mail..."
-echo "========================================================================================"	
+		file=$rootpath$thunderboltfile
+		if [ -f $file ]; then
+			echo "file $file - exist!"
+			AddOrReplaceParamInFile "# Disable thunderbolt devices" ""  $file
+			AddOrReplaceParamInFile "blacklist thunderbolt" ""  $file $FirstLevel
+			RemoveEmptyStrings $file
+		else
+			echo "file $file - NOT exist!!!!"
+		fi
+	echo ""
+fi
+if [ -n $mailfile ]; then
+	echo "Config Mail..."
+	echo "========================================================================================"	
 	file=$rootpath$mailfile
 	if [ -n $SMTPServer ]; then
 		if [ -f $file ]; then
-		 	true
+			true
 		else
 			touch $file
 		fi
@@ -184,25 +203,29 @@ echo "==========================================================================
 		sudo tail /var/log/mail.log
 		sleep 5s
 	fi
-echo ""
-echo "Config Syslog..."
-echo "========================================================================================"	
+	echo ""
+fi
+if [ -n $syslogfile ]; then
+	echo "Config Syslog..."
+	echo "========================================================================================"	
 	file=$rootpath$syslogfile
 	if [ -f $file ]; then
-        echo "file $file - exist!"
+		echo "file $file - exist!"
 		AddOrReplaceParamInFile "*.Warning @$Syslog:$SyslogPort" ""  $file		
 		RemoveEmptyStrings $file
 		sudo service rsyslog restart
 		#sudo service rsyslog status
 	else
-        echo "file $file - NOT exist!!!!"
-    fi	
-echo ""
-echo "Config Unattended-Upgrades..."
-echo "========================================================================================"
+		echo "file $file - NOT exist!!!!"
+	fi	
+	echo ""
+fi
+if [ -n $unattendedupgrades ]; then
+	echo "Config Unattended-Upgrades..."
+	echo "========================================================================================"
 	file=$rootpath$unattendedupgrades
 	if [ -f $file ]; then
-        echo "file $file - exist!"
+		echo "file $file - exist!"
 		AddOrReplaceParamInFile "Unattended-Upgrade::MinimalSteps " '"false";'  $file
 		AddOrReplaceParamInFile "Unattended-Upgrade::InstallOnShutdown " '"false";'  $file
 		AddOrReplaceParamInFile "Unattended-Upgrade::Mail " "\"$AdminMail\";"  $file
@@ -216,34 +239,37 @@ echo "==========================================================================
 		RemoveEmptyStrings $file
 		unattended-upgrade -v --dry-run
 	else
-        echo "file $file - NOT exist!!!!"
-    fi	
-echo ""
-sleep 2s
-echo "Remove boot delay..."
-echo "========================================================================================"
-	#sudo systemctl edit --full systemd-networkd-wait-online.service
-	#ExecStart=/lib/systemd/systemd-networkd-wait-online --timeout 1
-echo ""
+		echo "file $file - NOT exist!!!!"
+	fi	
 	sleep 2s
-echo "Install RKHunter IDS..."
-echo "========================================================================================"
+	echo ""	
+fi
+# echo "Remove boot delay..."
+# echo "========================================================================================"
+# 	#sudo systemctl edit --full systemd-networkd-wait-online.service
+# 	#ExecStart=/lib/systemd/systemd-networkd-wait-online --timeout 1
+# echo ""
+# 	sleep 2s
+if [ -n $rkhunterfile ]; then
+	echo "Install RKHunter IDS..."
+	echo "========================================================================================"
 	sudo apt-get install rkhunter -y
 	sleep 20s
 	file=$rootpath$rkhunterfile
 	if [ -f $file ]; then
-        echo "file $file - exist!"
+		echo "file $file - exist!"
 		AddOrReplaceParamInFile "UPDATE_MIRRORS=" "1"  $file $FirstLevel
 		AddOrReplaceParamInFile "MIRRORS_MODE=" "0"  $file $FirstLevel
 		AddOrReplaceParamInFile "/'WEB_CMD=" "/"/"/'"  $file $FirstLevel
 	else
-        echo "file $file - NOT exist!!!!"
-    fi
+		echo "file $file - NOT exist!!!!"
+	fi
 	#echo 'WEB_CMD=""' >> /etc/rkhunter.conf
 	sudo rkhunter --check --sk
 	sudo rkhunter --update
-echo ""
 	sleep 2s
+	echo ""	
+fi
 echo "Installing Fail2ban..."
 echo "========================================================================================"
 	sudo apt-get install fail2ban -y	
@@ -252,45 +278,50 @@ echo "Remove unused packages..."
 echo "========================================================================================"
 	sudo apt autoremove -y
 echo ""
-echo "Add scheduled tasks..."
-echo "========================================================================================"
+if [ -n $rkhunterfile ]; then
+	echo "Add scheduled tasks..."
+	echo "========================================================================================"
 	#write out current crontab
 	file="/tmp/mycron"
 	sudo crontab -l > $file
 	#echo new cron into cron file
-echo " - Reboot every sat at 3:00"   
-	AddOrReplaceParamInFile "0 3 * * 6 sudo sh -c '/bin/date>>/var/log/reboot.log && uptime>>/var/log/reboot.log && /sbin/reboot'" ""  $file $FirstLevel
-echo " - Update rkhunter every night at 2:30"     
-	AddOrReplaceParamInFile "30 2 * * * sudo sh -c '/usr/bin/rkhunter --update>>/var/log/rkhunter-upd.log'" ""  $file $FirstLevel
-echo " - Run rkhunter every night at 2:40"     
-	AddOrReplaceParamInFile "40 2 * * * sudo sh -c '/usr/bin/rkhunter'" ""  $file $FirstLevel
-echo " - Send rkhunter logs every night at 2:55"
-	AddOrReplaceParamInFile "55 2 * * 7 grep -e 'Warning' -e 'Error' -e 'Fail' /var/log/rkhunter.log | mail -s '$Company RKHunter' -a 'From: RKHunter@$HostName.$Company.local' $AdminMail" ""  $file $FirstLevel    
+	echo "$Task1Comment"   
+	AddOrReplaceParamInFile "$Task1time $Task1" ""  $file $FirstLevel
+	echo "$Task2Comment"     
+	AddOrReplaceParamInFile "$Task2time $Task2" ""  $file $FirstLevel
+	echo "$Task2Comment"     
+	AddOrReplaceParamInFile "$Task3time $Task3" ""  $file $FirstLevel
+	echo "$Task4Comment"
+	AddOrReplaceParamInFile "$Task4time $Task4" ""  $file $FirstLevel    
 	#install new cron file
 	sudo crontab $file
 	sudo rm $file
+fi
+if [ -n $sshdfile ]; then
+	echo "Check config files..."
+	echo "========================================================================================"
+	CheckResult="/opt/projects/scripts/CheckResult.txt"
 
-echo "Check config files..."
-echo "========================================================================================"
-CheckResult="/opt/projects/scripts/CheckResult.txt"
+	file=$rootpath$sshdfile
+	CorrectAction="service sshd restart"
+	CheckConfigFile "sshd -t -f $file" $file $CheckResult "$CorrectAction"
 
-file=$rootpath$sshdfile
-CorrectAction="service sshd restart"
-CheckConfigFile "sshd -t -f $file" $file $CheckResult "$CorrectAction"
+	echo ""
+	file=$rootpath$sysctlfile
+	CheckConfigFile "sysctl -p $file | grep 'sysctl:'" $file $CheckResult
 
-echo ""
-file=$rootpath$sysctlfile
-CheckConfigFile "sysctl -p $file | grep 'sysctl:'" $file $CheckResult
-
-rm $CheckResult
-
+	rm $CheckResult
+fi
 echo ""
 echo "Script complited!"
 read -p "Reboot the system now? [Y/N]" answer
 case $answer in
-   [yY]* ) sudo reboot;;
+   [yY]* ) sudo reboot 
+   ;;
 
-   [nN]* ) exit;;
+   [nN]* ) exit
+   ;;
 
-   * )     exit;;
+   * )     exit
+   ;;
 esac
